TEST: Reducing duplication in integration tests
By adding a setup method in the base class which automatically
compiles the test code and starts the Zemu instance,
duplication in individual integration test classes has been
reduced.

TEST: Removing calls to compile and start instance

Removing calls to compile_test_code() and start_instance()
in integration tests, with the plan to do this instead
in the test method initializer

TEST: Compiling test code in setup() method

The integration test base class now defines a setup()
method which compiles the test code for each test and
starts the emulator instance.

TEST: Fixing test_boot

TEST: Renamed file to fit with new test structure

TEST: Moved test source code into separate folders

TEST: Moved assembly source code for tests

TEST: Changed paths in base class for tests

TEST: Sorting out paths to user code used in tests

TEST: Refactoring how user programs are loaded

TEST: Changing where compiled test files get generated

TEST: Adding helper function for loading map files